+++ packages/react-openapi-client-generator/examples/todoapp/src/mocks/browser.ts
@@ -1,5 +1,5 @@
 import { setupWorker } from 'msw';
-import { handlers } from '@visma/msw-openapi-backend-integration';
+import { handlers } from '@twoday/msw-openapi-backend-integration';
 import requestLogicHandlers, { definition } from './requestLogicHandlers';
 
 export const worker = setupWorker(