+++ packages/tailwindcss-vud/vanilla-theme.js
@@ -1,3 +1,0 @@
-module.exports = {
-  // to be implemented
-};