+++ packages/vite-plugin-super-template/src/dynamicBase.ts
@@ -1,10 +1,10 @@
-import '@visma/public.config/config';
+import '@twoday/public.config/config';
 import { Plugin } from 'vite';
 import { dynamicBase as vitePluginDynamicBase } from 'vite-plugin-dynamic-base';
 
 const dynamicBase: Plugin[] = [
   {
-    name: '@visma/vite-plugin-dynamic-base',
+    name: '@twoday/vite-plugin-dynamic-base',
     config: () => ({
       base: process.env.NODE_ENV === 'production' ? '/__dynamic_base__/' : '/',
     }),