+++ packages/react-keycloak/CHANGELOG.md
@@ -10,7 +10,7 @@
 
 # 1.0.0
 
-- **Breaking:** Upgrade to `@visma/public.config@1.0.0`
+- **Breaking:** Upgrade to `@twoday/public.config@1.0.0`
 
 # 0.1.3